requirements.txt
tk
pillow
# csv
awesometkinter
auto-py-to-exe
